# microservices/gatewayService/docs/apifox-gateway-openapi.yaml
openapi: 3.0.3
info:
  title: Sky Takeout Gateway API
  version: 1.0.0
  description: |
    Gateway Service API definitions generated from current gin routes.
    This file can be imported directly into Apifox.
servers:
  - url: http://localhost:18080
    description: Local docker runtime
...
paths:
...
  /getway/token/exchange:
    post:
      tags: [Auth]
      summary: Exchange username/password for token
      operationId: tokenExchange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenExchangeRequest'
            examples:
              default:
                value:
                  username: admin
                  password: "123456"
      responses:
        '200':
          description: Unified business response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'
              examples:
                success:
                  value:
                    code: 1
                    msg: ok
                    data:
                      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      username: admin
                invalidParams:
                  value:
                    code: 403
                    msg: "参数错误: Key: 'tokenExchangeRequest.Username' Error:Field validation for 'Username' failed on the 'required' tag"
                    data: null
                authFailed:
                  value:
                    code: 403
                    msg: "认证失败: 账号不存在"
                    data: null
  /getway/token/refresh:
    post:
      tags: [Auth]
      summary: Refresh token (validate refreshToken first)
      operationId: tokenRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
            examples:
              default:
                value:
                  refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  username: admin
                  password: "123456"
      responses:
        '200':
          description: Unified business response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'
              examples:
                success:
                  value:
                    code: 1
                    msg: ok
                    data:
                      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      username: admin
                      expireDate: "2026-06-03 15:04:05 +0800 CST"
                invalidToken:
                  value:
                    code: 403
                    msg: "无效的 refreshToken"
                    data: null
...
components:
  schemas:
...
    TokenExchangeRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: 123456
    TokenRefreshRequest:
      type: object
      required: [refreshToken, username, password]
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        username:
          type: string
          example: admin
        password:
          type: string
          example: 123456
    UnifiedResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          description: Business code from common/e/code.go
          example: 1
        msg:
          type: string
          example: ok
        data:
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/TokenExchangeData'
            - $ref: '#/components/schemas/TokenRefreshData'
            - type: object
              additionalProperties: true
    TokenExchangeData:
      type: object
      required: [token, username]
      properties:
        token:
          type: string
        username:
          type: string
    TokenRefreshData:
      type: object
      required: [token, username, expireDate]
      properties:
        token:
          type: string
        username:
          type: string
        expireDate:
          type: string
          description: Local time string generated by Go time formatting
...